refactor : bottom bar로 이름 변경 및 navigation bar 위치 변경 #49
